MVC2 arcade support, new SFA3 Nash stage

+++ palmod/Game/Game_MSHVSF_A.cpp
@@ -649,7 +649,7 @@ void CGame_MSHVSF_A::LoadSpecificPaletteData(UINT16 nUnitId, UINT16 nPalId)
         int cbPaletteSizeOnDisc = 0;
         const sGame_PaletteDataset* paletteData = GetSpecificPalette(nUnitId, nPalId);
 
-        cbPaletteSizeOnDisc = max(0, (paletteData->nPaletteOffsetEnd - paletteData->nPaletteOffset));
+        cbPaletteSizeOnDisc = (int)max(0, (paletteData->nPaletteOffsetEnd - paletteData->nPaletteOffset));
 
         nCurrPalOffs = paletteData->nPaletteOffset;
         m_nCurrentPaletteSize = cbPaletteSizeOnDisc / 2;
