+++ palmod/Game/Game_SSF2T_A.cpp
@@ -621,7 +621,7 @@ void CGame_SSF2T_A::LoadSpecificPaletteData(UINT16 nUnitId, UINT16 nPalId)
         int cbPaletteSizeOnDisc = 0;
         const sGame_PaletteDataset* paletteData = GetSpecificPalette(nUnitId, nPalId);
 
-        cbPaletteSizeOnDisc = max(0, (paletteData->nPaletteOffsetEnd - paletteData->nPaletteOffset));
+        cbPaletteSizeOnDisc = (int)max(0, (paletteData->nPaletteOffsetEnd - paletteData->nPaletteOffset));
 
         nCurrPalOffs = paletteData->nPaletteOffset;
         m_nCurrentPaletteSize = cbPaletteSizeOnDisc / 2;
